backend/requirements/test.txt
black==19.10b0
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
pytest-asyncio==0.10.0
factory_boy==2.11.1
flake8==3.7.7
flake8-isort==2.7.0
coverage==4.5.3
moto==1.3.7